project/requirements.txt
coverage==4.5.2
Django==2.1.5
djangorestframework==3.9.1
pkg-resources==0.0.0
pytz==2018.9
